+++ src/main/java/com/NetCracker/Exceptions/AppointmentNotFoundException.java
@@ -1,7 +1,7 @@
 package com.NetCracker.Exceptions;
 
 public class AppointmentNotFoundException extends RuntimeException {
-    public AppointmentNotFoundException(Integer id) {
+    public AppointmentNotFoundException(Long id) {
         super("Could not find an appointment " + id);
     }
 }
