+++ src/main/java/com/NetCracker/Exceptions/AppointmentRegistrationNotFoundException.java
@@ -1,7 +1,7 @@
 package com.NetCracker.Exceptions;
 
 public class AppointmentRegistrationNotFoundException extends RuntimeException {
-    public AppointmentRegistrationNotFoundException(Integer id) {
+    public AppointmentRegistrationNotFoundException(Long id) {
         super("Could not find an appointment registration " + id);
     }
 }
